--- ipaas/src/pages/McpPlaygroundInspector.tsx
@@ -1,160 +1,0 @@
-/**
- * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { Alert, Box, MenuItem, PageContent, Select } from '@wso2/oxygen-ui';
-import { useEffect, useMemo, useState, type JSX } from 'react';
-import ComingSoon from './ComingSoon';
-import McpPlayground from '../components/McpPlayground/McpPlayground';
-import DeploymentTrackBar from '../components/DeploymentTrackBar';
-import { useGenerateTestKey } from '../hooks/useApim';
-import { useComponentByHandler } from '../hooks/useComponents';
-import { useComponentDeployment, useEnvEndpoints } from '../hooks/useDeployments';
-import { useEnvironments } from '../hooks/useEnvironments';
-import { useOrgUuid } from '../hooks/useOrgUuid';
-import { useProjectId } from '../hooks/useProjects';
-import type { ComponentScope } from '../nav';
-import type { EnvEndpoint } from '../types/component';
-
-const TEST_KEY_HEADER = 'test-key';
-
-/** Network-visibility URL resolvers (mirrors the endpoint URLs panel). */
-const VISIBILITY_OPTIONS: { value: string; label: string; getUrl: (ep: EnvEndpoint) => string }[] = [
-  { value: 'public', label: 'Public', getUrl: (ep) => ep.publicUrl || ep.defaultPublicUrl || ep.invokeUrl || '' },
-  { value: 'organization', label: 'Organization', getUrl: (ep) => ep.organizationUrl || ep.defaultOrganizationUrl || '' },
-  { value: 'project', label: 'Project', getUrl: (ep) => ep.projectUrl || '' },
-];
-
-/**
- * Resolves the deployed MCP endpoint + a freshly minted `test-key` for the selected
- * environment/track, then hands them to the in-house {@link McpPlayground} (the interactive
- * connect / list-tools / invoke / ping surface).
- */
-export default function McpPlaygroundInspector({ scope }: { scope: ComponentScope }): JSX.Element {
-  const orgUuid = useOrgUuid() ?? '';
-  const { projectId, project } = useProjectId(scope.project);
-  const { data: component } = useComponentByHandler(projectId, scope.component);
-  const isMcp = component?.componentSubType === 'MCP';
-
-  const { data: environments = [] } = useEnvironments(scope.org, projectId);
-
-  // Deployment track (default latest).
-  const tracks = useMemo(() => component?.deploymentTracks ?? [], [component?.deploymentTracks]);
-  const [selectedTrackId, setSelectedTrackId] = useState('');
-  useEffect(() => {
-    if (!tracks.length) return;
-    setSelectedTrackId((prev) => (prev && tracks.some((t) => t.id === prev) ? prev : (tracks.find((t) => t.latest)?.id ?? tracks[0].id)));
-  }, [component?.id, tracks]);
-
-  // Environment selection.
-  const [selectedEnvId, setSelectedEnvId] = useState('');
-  useEffect(() => {
-    if (!environments.length) return;
-    setSelectedEnvId((prev) => (prev && environments.some((e) => e.id === prev) ? prev : environments[0].id));
-  }, [environments]);
-  const selectedEnv = environments.find((e) => e.id === selectedEnvId) ?? null;
-
-  // Deployment → releaseId → endpoints for the selected env + track.
-  const { data: deployment } = useComponentDeployment(component ? scope.org : '', component ? orgUuid : '', component?.id ?? '', selectedTrackId, selectedEnv?.id ?? '');
-  const releaseId = deployment?.releaseId ?? '';
-  const { data: endpoints = [] } = useEnvEndpoints(component?.id ?? '', selectedTrackId, releaseId);
-
-  // Testable endpoints (those with an APIM id). The user picks the endpoint + visibility.
-  const testableEndpoints = useMemo(() => endpoints.filter((e) => e.apimId), [endpoints]);
-  const [selectedEndpointId, setSelectedEndpointId] = useState('');
-  const [selectedVisibility, setSelectedVisibility] = useState('');
-
-  const activeEndpointId = testableEndpoints.some((e) => e.id === selectedEndpointId) ? selectedEndpointId : (testableEndpoints[0]?.id ?? '');
-  const activeEndpoint = testableEndpoints.find((e) => e.id === activeEndpointId) ?? null;
-
-  const visibilityOptions = useMemo(() => (activeEndpoint ? VISIBILITY_OPTIONS.filter((v) => v.getUrl(activeEndpoint) && (!activeEndpoint.networkVisibilities?.length || activeEndpoint.networkVisibilities.includes(v.label))) : []), [activeEndpoint]);
-  const activeVisibility = visibilityOptions.some((v) => v.value === selectedVisibility) ? selectedVisibility : (visibilityOptions[0]?.value ?? '');
-
-  const baseUrl = visibilityOptions.find((v) => v.value === activeVisibility)?.getUrl(activeEndpoint!) ?? '';
-  const mcpUrl = baseUrl ? `${baseUrl}/mcp` : '';
-  const apimId = activeEndpoint?.apimId ?? null;
-  const isDeploymentInProgress = deployment?.deploymentStatusV2 === 'IN_PROGRESS';
-
-  const endpointSwitcher = { options: testableEndpoints.map((e) => ({ label: e.displayName, value: e.id })), value: activeEndpointId, onChange: setSelectedEndpointId };
-  const visibilitySwitcher = { options: visibilityOptions.map((v) => ({ label: v.label, value: v.value })), value: activeVisibility, onChange: setSelectedVisibility };
-
-  // Mint a test key for the selected environment (sent as the `test-key` header).
-  const generateKey = useGenerateTestKey();
-  const [token, setToken] = useState('');
-  const [tokenFetching, setTokenFetching] = useState(false);
-  const [regenNonce, setRegenNonce] = useState(0);
-  useEffect(() => {
-    if (!apimId) {
-      setToken('');
-      return undefined;
-    }
-    let cancelled = false;
-    setTokenFetching(true);
-    generateKey
-      .mutateAsync({ apimId, keyType: selectedEnv?.critical ? 'Production' : 'Development' })
-      .then((r) => {
-        if (!cancelled) setToken(r?.apikey ?? '');
-      })
-      .catch(() => {
-        if (!cancelled) setToken('');
-      })
-      .finally(() => {
-        if (!cancelled) setTokenFetching(false);
-      });
-    return () => {
-      cancelled = true;
-    };
-    // generateKey is a stable mutation; apimId/env/regen drive identity
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [apimId, selectedEnv?.critical, regenNonce]);
-
-  // Non-MCP components keep this route's previous Coming Soon behaviour.
-  if (component && !isMcp) {
-    return <ComingSoon title="Coming Soon" description="Testing tools are currently under development." />;
-  }
-
-  const envSelector = environments.length > 0 && (
-    <Select
-      size="small"
-      value={selectedEnvId}
-      onChange={(e) => setSelectedEnvId(e.target.value as string)}
-      inputProps={{ 'aria-label': 'Environment' }}
-      sx={{ fontSize: '0.8125rem', '& .MuiOutlinedInput-notchedOutline': { borderRadius: 5 }, '& .MuiSelect-select': { py: 0.5, px: 1.5 }, minWidth: 140 }}>
-      {environments.map((env) => (
-        <MenuItem key={env.id} value={env.id}>
-          {env.name}
-        </MenuItem>
-      ))}
-    </Select>
-  );
-
-  return (
-    <Box sx={{ display: 'flex', flexDirection: 'column', height: '100%', minHeight: 0 }}>
-      {tracks.length > 0 && <DeploymentTrackBar tracks={tracks} selectedId={selectedTrackId} onChange={setSelectedTrackId} orgHandler={scope.org} projectHandler={project?.handler ?? scope.project} componentHandler={scope.component} extra={envSelector} />}
-
-      <PageContent sx={{ flex: 1, minHeight: 0, display: 'flex', flexDirection: 'column' }}>
-        {isDeploymentInProgress ? (
-          <Alert severity="info">Integration deployment is still in progress. You can start testing when the deployment is complete.</Alert>
-        ) : mcpUrl ? (
-          <McpPlayground url={mcpUrl} token={token || null} headerName={TEST_KEY_HEADER} isTokenFetching={tokenFetching} onTokenRegenerate={() => setRegenNonce((n) => n + 1)} endpointSwitcher={endpointSwitcher} visibilitySwitcher={visibilitySwitcher} />
-        ) : (
-          <Alert severity="warning">The integration is still not deployed. You must deploy the integration to test it.</Alert>
-        )}
-      </PageContent>
-    </Box>
-  );
-}
